requirements.txt
asgiref==3.7.2
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
flask==2.3.3
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
Pillow==10.1.0
PyJWT==1.7.1
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
werkzeug==2.3.7
zipp==3.17.0
